generate devcontainer feature source code from lsp-bridge

+++ _generator/generator.py
@@ -5,35 +5,38 @@ import os
 from cookiecutter.main import cookiecutter
 
 # Read the JSON file
-with open('langserver.json', 'r') as file:
+with open("langserver.json", "r") as file:
     data = json.load(file)
 
 template_src_dir = "src_template"
 template_test_dir = "test_template"
 
 for entry in data:
-    langserver = entry['langserver']
+    if not entry["packages"]:
+        continue
+
+    langserver = entry["langserver"]
     target_src_dir = f"../src/{langserver}"
     target_test_dir = f"../test/{langserver}"
 
-    if os.path.exists(target_src_dir):    
-        shutil.rmtree(target_src_dir)        
+    if os.path.exists(target_src_dir):
+        shutil.rmtree(target_src_dir)
 
-    if os.path.exists(target_test_dir):     
+    if os.path.exists(target_test_dir):
         shutil.rmtree(target_test_dir)
- 
+
     cookiecutter(
-        'src_template',
-        no_input=True,     
+        "src_template",
+        no_input=True,
         output_dir=f"{target_src_dir}/../",
-        extra_context=entry
-    )   
+        extra_context=entry,
+    )
     print(f"Generated {target_src_dir}")
 
     cookiecutter(
-        'test_template',
-        no_input=True,     
+        "test_template",
+        no_input=True,
         output_dir=f"{target_test_dir}/../",
-        extra_context=entry
+        extra_context=entry,
     )
     print(f"Generated {target_test_dir}")
